distance removed from single discount

+++ app/src/main/java/com/roadway/capslabs/roadway_chat/models/Event.java
@@ -12,7 +12,7 @@ public class Event {
     private final String address;
     private final String metro;
     private final DateRange range;
-    private final float rating;
+    private final int rating;
     private final int id;
     private final double distance;
     private final String pictureUrl;
@@ -24,7 +24,7 @@ public class Event {
             title = eventObj.getString("title");
             description = eventObj.getString("about");
             range = new DateRange(eventObj.getString("date_start"), eventObj.getString("date_end"));
-            rating = (float) eventObj.getDouble("rating");
+            rating = (int) eventObj.getDouble("rating");
             id = eventObj.getInt("id");
             pictureUrl = eventObj.getString("avatar");
             address = eventObj.getString("address");
@@ -56,7 +56,7 @@ public class Event {
         return range.getDateEnd();
     }
 
-    public float getRating() {
+    public int getRating() {
         return rating;
     }
